+++ src/chunk_append/exec.c
@@ -6,16 +6,22 @@
 
 #include <postgres.h>
 #include <miscadmin.h>
+#include <executor/executor.h>
+#include <executor/nodeSubplan.h>
 #include <nodes/bitmapset.h>
 #include <nodes/makefuncs.h>
 #include <nodes/nodeFuncs.h>
+#include <nodes/relation.h>
 #include <optimizer/clauses.h>
+#include <optimizer/cost.h>
 #include <optimizer/plancat.h>
+#include <optimizer/predtest.h>
+#include <optimizer/prep.h>
 #include <optimizer/restrictinfo.h>
 #include <parser/parsetree.h>
 #include <rewrite/rewriteManip.h>
-#include <utils/typcache.h>
 #include <utils/memutils.h>
+#include <utils/typcache.h>
 
 #include "chunk_append/chunk_append.h"
 #include "chunk_append/exec.h"
@@ -37,13 +43,12 @@ static CustomExecMethods chunk_append_state_methods = {
 };
 
 static List *constify_restrictinfos(PlannerInfo *root, List *restrictinfos);
-static bool can_exclude_chunk(PlannerInfo *root, EState *estate, Index rt_index,
-							  List *restrictinfos);
+static bool can_exclude_chunk(List *constraints, List *restrictinfos);
 static void do_startup_exclusion(ChunkAppendState *state);
 static Node *constify_param_mutator(Node *node, void *context);
 static List *constify_restrictinfo_params(PlannerInfo *root, EState *state, List *restrictinfos);
 
-static void adjust_ri_clauses(ChunkAppendState *state, List *initial_rt_indexes);
+static void initialize_constraints(ChunkAppendState *state, List *initial_rt_indexes);
 
 Node *
 chunk_append_state_create(CustomScan *cscan)
@@ -54,20 +59,14 @@ chunk_append_state_create(CustomScan *cscan)
 
 	state->csstate.methods = &chunk_append_state_methods;
 
-	state->num_subplans = 0;
-	state->subplanstates = NULL;
-
 	state->initial_subplans = cscan->custom_plans;
 	state->initial_ri_clauses = lsecond(cscan->custom_private);
-	adjust_ri_clauses(state, lthird(cscan->custom_private));
 	state->sort_options = lfourth(cscan->custom_private);
 
 	state->startup_exclusion = (bool) linitial_oid(linitial(cscan->custom_private));
 	state->runtime_exclusion = (bool) lsecond_oid(linitial(cscan->custom_private));
 	state->limit = lthird_oid(linitial(cscan->custom_private));
 
-	state->current = 0;
-	state->runtime_initialized = false;
 	state->filtered_subplans = state->initial_subplans;
 	state->filtered_ri_clauses = state->initial_ri_clauses;
 
@@ -79,30 +78,33 @@ do_startup_exclusion(ChunkAppendState *state)
 {
 	List *filtered_children = NIL;
 	List *filtered_ri_clauses = NIL;
+	List *filtered_constraints = NIL;
 	ListCell *lc_plan;
 	ListCell *lc_clauses;
+	ListCell *lc_constraints;
 
 	/*
-	 * create skeleton plannerinfo to reuse some PostgreSQL planner functions
+	 * create skeleton plannerinfo for estimate_expression_value
 	 */
-	Query parse = {
-		.resultRelation = InvalidOid,
-	};
 	PlannerGlobal glob = {
 		.boundParams = NULL,
 	};
 	PlannerInfo root = {
 		.glob = &glob,
-		.parse = &parse,
 	};
 
 	/*
-	 * clauses should always have the same length as appendplans because
-	 * the list of clauses is built from the list of appendplans
+	 * clauses and constraints should always have the same length as initial_subplans
 	 */
 	Assert(list_length(state->initial_subplans) == list_length(state->initial_ri_clauses));
+	Assert(list_length(state->initial_subplans) == list_length(state->initial_constraints));
 
-	forboth (lc_plan, state->initial_subplans, lc_clauses, state->initial_ri_clauses)
+	forthree (lc_plan,
+			  state->initial_subplans,
+			  lc_constraints,
+			  state->initial_constraints,
+			  lc_clauses,
+			  state->initial_ri_clauses)
 	{
 		List *restrictinfos = NIL;
 		List *ri_clauses = lfirst(lc_clauses);
@@ -123,18 +125,18 @@ do_startup_exclusion(ChunkAppendState *state)
 			}
 			restrictinfos = constify_restrictinfos(&root, restrictinfos);
 
-			if (can_exclude_chunk(&root,
-								  state->csstate.ss.ps.state,
-								  scan->scanrelid,
-								  restrictinfos))
+			if (can_exclude_chunk(lfirst(lc_constraints), restrictinfos))
 				continue;
 		}
 
 		filtered_children = lappend(filtered_children, lfirst(lc_plan));
 		filtered_ri_clauses = lappend(filtered_ri_clauses, ri_clauses);
+		filtered_constraints = lappend(filtered_constraints, lfirst(lc_constraints));
 	}
+
 	state->filtered_subplans = filtered_children;
 	state->filtered_ri_clauses = filtered_ri_clauses;
+	state->filtered_constraints = filtered_constraints;
 }
 
 static void
@@ -144,6 +146,9 @@ chunk_append_begin(CustomScanState *node, EState *estate, int eflags)
 	ChunkAppendState *state = (ChunkAppendState *) node;
 	ListCell *lc;
 	int i;
+
+	Assert(list_length(cscan->custom_plans) == list_length(state->initial_subplans));
+	initialize_constraints(state, lthird(cscan->custom_private));
 
 	if (state->startup_exclusion)
 		do_startup_exclusion(state);
@@ -175,29 +180,42 @@ chunk_append_begin(CustomScanState *node, EState *estate, int eflags)
 
 		i++;
 	}
+
+	if (state->runtime_exclusion)
+	{
+		state->params = state->subplanstates[0]->plan->allParam;
+		/*
+		 * make sure all params are initialized for runtime exclusion
+		 */
+		node->ss.ps.chgParam = state->subplanstates[0]->plan->allParam;
+	}
 }
 
+/*
+ * build bitmap of valid subplans for runtime exclusion
+ */
 static void
 initialize_runtime_exclusion(ChunkAppendState *state)
 {
-	ListCell *lc_clauses;
+	ListCell *lc_clauses, *lc_constraints;
 	int i = 0;
 
-	Query parse = {
-		.resultRelation = InvalidOid,
-	};
 	PlannerGlobal glob = {
 		.boundParams = NULL,
 	};
 	PlannerInfo root = {
 		.glob = &glob,
-		.parse = &parse,
 	};
 
 	Assert(state->num_subplans == list_length(state->filtered_ri_clauses));
 
 	lc_clauses = list_head(state->filtered_ri_clauses);
+	lc_constraints = list_head(state->filtered_constraints);
 
+	state->runtime_number_loops++;
+	/*
+	 * mark subplans as active/inactive in valid_subplans
+	 */
 	for (i = 0; i < state->num_subplans; i++)
 	{
 		PlanState *ps = state->subplanstates[i];
@@ -219,11 +237,14 @@ initialize_runtime_exclusion(ChunkAppendState *state)
 			}
 			restrictinfos = constify_restrictinfo_params(&root, ps->state, restrictinfos);
 
-			if (!can_exclude_chunk(&root, ps->state, scan->scanrelid, restrictinfos))
+			if (!can_exclude_chunk(lfirst(lc_constraints), restrictinfos))
 				state->valid_subplans = bms_add_member(state->valid_subplans, i);
+			else
+				state->runtime_number_exclusions++;
 		}
 
 		lc_clauses = lnext(lc_clauses);
+		lc_constraints = lnext(lc_constraints);
 	}
 
 	state->runtime_initialized = true;
@@ -251,7 +272,7 @@ chunk_append_exec(CustomScanState *node)
 	{
 		initialize_runtime_exclusion(state);
 
-		if (!state->valid_subplans || bms_num_members(state->valid_subplans) == 0)
+		if (bms_is_empty(state->valid_subplans))
 			return ExecClearTuple(node->ss.ps.ps_ResultTupleSlot);
 
 		state->current = bms_next_member(state->valid_subplans, -1);
@@ -288,9 +309,8 @@ chunk_append_exec(CustomScanState *node)
 		if (!TupIsNull(subslot))
 		{
 			/*
-			 * If the subplan gave us something then return it as-is. We do
-			 * NOT make use of the result slot that was set up in
-			 * chunk_append_begin there's no need for it.
+			 * If the subplan gave us something check if we need
+			 * to do projection otherwise return as is.
 			 */
 			if (node->ss.ps.ps_ProjInfo == NULL)
 				return subslot;
@@ -348,11 +368,17 @@ chunk_append_rescan(CustomScanState *node)
 
 	for (i = 0; i < state->num_subplans; i++)
 	{
+		if (node->ss.ps.chgParam != NULL)
+			UpdateChangedParamSet(state->subplanstates[i], node->ss.ps.chgParam);
+
 		ExecReScan(state->subplanstates[i]);
 	}
 	state->current = 0;
 
-	if (state->runtime_exclusion)
+	/*
+	 * detect changed params and reset runtime exclusion state
+	 */
+	if (state->runtime_exclusion && bms_overlap(node->ss.ps.chgParam, state->params))
 	{
 		bms_free(state->valid_subplans);
 		state->valid_subplans = NULL;
@@ -411,20 +437,26 @@ constify_param_mutator(Node *node, void *context)
 	if (IsA(node, Param))
 	{
 		Param *param = castNode(Param, node);
-		EState *state = (EState *) context;
+		EState *estate = (EState *) context;
 
 		if (param->paramkind == PARAM_EXEC)
 		{
 			TypeCacheEntry *tce = lookup_type_cache(param->paramtype, 0);
-			ParamExecData value = state->es_param_exec_vals[param->paramid];
+			ParamExecData prm = estate->es_param_exec_vals[param->paramid];
 
-			if (!value.execPlan)
+			if (prm.execPlan != NULL)
+			{
+				ExprContext *econtext = GetPerTupleExprContext(estate);
+				ExecSetParamPlan(prm.execPlan, econtext);
+			}
+
+			if (prm.execPlan == NULL)
 				return (Node *) makeConst(param->paramtype,
 										  param->paramtypmod,
 										  param->paramcollid,
 										  tce->typlen,
-										  value.value,
-										  value.isnull,
+										  prm.value,
+										  prm.isnull,
 										  tce->typbyval);
 		}
 		return node;
@@ -434,37 +466,166 @@ constify_param_mutator(Node *node, void *context)
 }
 
 /*
- * Exclude child relations (chunks) at execution time based on constraints.
- *
- * This functions tries to reuse as much functionality as possible from standard
- * constraint exclusion in PostgreSQL that normally happens at planning
- * time. Therefore, we need to fake a number of planning-related data
- * structures.
+ * stripped down version of postgres get_relation_constraints
  */
-static bool
-can_exclude_chunk(PlannerInfo *root, EState *estate, Index rt_index, List *restrictinfos)
+static List *
+ca_get_relation_constraints(Oid relationObjectId, Index varno, bool include_notnull)
 {
-	RangeTblEntry *rte = rt_fetch(rt_index, estate->es_range_table);
-	RelOptInfo rel = {
-		.type = T_RelOptInfo,
-		.relid = rt_index,
-		.reloptkind = RELOPT_OTHER_MEMBER_REL,
-		.baserestrictinfo = restrictinfos,
-	};
+	List *result = NIL;
+	Relation relation;
+	TupleConstr *constr;
 
-	return rte->rtekind == RTE_RELATION && rte->relkind == RELKIND_RELATION && !rte->inh &&
-		   relation_excluded_by_constraints(root, &rel, rte);
+	/*
+	 * We assume the relation has already been safely locked.
+	 */
+	relation = heap_open(relationObjectId, NoLock);
+
+	constr = relation->rd_att->constr;
+	if (constr != NULL)
+	{
+		int num_check = constr->num_check;
+		int i;
+
+		for (i = 0; i < num_check; i++)
+		{
+			Node *cexpr;
+
+			/*
+			 * If this constraint hasn't been fully validated yet, we must
+			 * ignore it here.
+			 */
+			if (!constr->check[i].ccvalid)
+				continue;
+
+			cexpr = stringToNode(constr->check[i].ccbin);
+
+			/*
+			 * Run each expression through const-simplification and
+			 * canonicalization.  This is not just an optimization, but is
+			 * necessary, because we will be comparing it to
+			 * similarly-processed qual clauses, and may fail to detect valid
+			 * matches without this.  This must match the processing done to
+			 * qual clauses in preprocess_expression()!  (We can skip the
+			 * stuff involving subqueries, however, since we don't allow any
+			 * in check constraints.)
+			 */
+			cexpr = eval_const_expressions(NULL, cexpr);
+
+#if (PG96 && PG_VERSION_NUM < 90609) || (PG10 && PG_VERSION_NUM < 100004)
+			cexpr = (Node *) canonicalize_qual((Expr *) cexpr);
+#elif PG96 || PG10
+			cexpr = (Node *) canonicalize_qual_ext((Expr *) cexpr, true);
+#else
+			cexpr = (Node *) canonicalize_qual((Expr *) cexpr, true);
+#endif
+
+			/* Fix Vars to have the desired varno */
+			if (varno != 1)
+				ChangeVarNodes(cexpr, 1, varno, 0);
+
+			/*
+			 * Finally, convert to implicit-AND format (that is, a List) and
+			 * append the resulting item(s) to our output list.
+			 */
+			result = list_concat(result, make_ands_implicit((Expr *) cexpr));
+		}
+
+		/* Add NOT NULL constraints in expression form, if requested */
+		if (include_notnull && constr->has_not_null)
+		{
+			int natts = relation->rd_att->natts;
+
+			for (i = 1; i <= natts; i++)
+			{
+				Form_pg_attribute att = TupleDescAttr(relation->rd_att, i - 1);
+
+				if (att->attnotnull && !att->attisdropped)
+				{
+					NullTest *ntest = makeNode(NullTest);
+
+					ntest->arg = (Expr *)
+						makeVar(varno, i, att->atttypid, att->atttypmod, att->attcollation, 0);
+					ntest->nulltesttype = IS_NOT_NULL;
+
+					/*
+					 * argisrow=false is correct even for a composite column,
+					 * because attnotnull does not represent a SQL-spec IS NOT
+					 * NULL test in such a case, just IS DISTINCT FROM NULL.
+					 */
+					ntest->argisrow = false;
+					ntest->location = -1;
+					result = lappend(result, ntest);
+				}
+			}
+		}
+	}
+
+	heap_close(relation, NoLock);
+
+	return result;
 }
 
 /*
- * Adjust the RangeTableEntry indexes in the restrictinfo
- * clauses because during planning subquery indexes will be
- * different from the final index after flattening.
+ * Exclude child relations (chunks) at execution time based on constraints.
+ *
+ * constraints is the list of constraint expressions of the relation
+ * baserestrictinfo is the list of RestrictInfos
+ */
+static bool
+can_exclude_chunk(List *constraints, List *baserestrictinfo)
+{
+	/*
+	 * Regardless of the setting of constraint_exclusion, detect
+	 * constant-FALSE-or-NULL restriction clauses.  Because const-folding will
+	 * reduce "anything AND FALSE" to just "FALSE", any such case should
+	 * result in exactly one baserestrictinfo entry.  This doesn't fire very
+	 * often, but it seems cheap enough to be worth doing anyway.  (Without
+	 * this, we'd miss some optimizations that 9.5 and earlier found via much
+	 * more roundabout methods.)
+	 */
+	if (list_length(baserestrictinfo) == 1)
+	{
+		RestrictInfo *rinfo = (RestrictInfo *) linitial(baserestrictinfo);
+		Expr *clause = rinfo->clause;
+
+		if (clause && IsA(clause, Const) &&
+			(((Const *) clause)->constisnull || !DatumGetBool(((Const *) clause)->constvalue)))
+			return true;
+	}
+
+	/*
+	 * The constraints are effectively ANDed together, so we can just try to
+	 * refute the entire collection at once.  This may allow us to make proofs
+	 * that would fail if we took them individually.
+	 *
+	 * Note: we use rel->baserestrictinfo, not safe_restrictions as might seem
+	 * an obvious optimization.  Some of the clauses might be OR clauses that
+	 * have volatile and nonvolatile subclauses, and it's OK to make
+	 * deductions with the nonvolatile parts.
+	 *
+	 * We need strong refutation because we have to prove that the constraints
+	 * would yield false, not just NULL.
+	 */
+#if PG96
+	if (predicate_refuted_by(constraints, baserestrictinfo))
+#else
+	if (predicate_refuted_by(constraints, baserestrictinfo, false))
+#endif
+		return true;
+
+	return false;
+}
+
+/*
+ * Fetch the constraints for a relation and adjust range table indexes
+ * if necessary.
  */
 static void
-adjust_ri_clauses(ChunkAppendState *state, List *initial_rt_indexes)
+initialize_constraints(ChunkAppendState *state, List *initial_rt_indexes)
 {
 	ListCell *lc_clauses, *lc_plan, *lc_relid;
+	List *constraints = NIL;
+	EState *estate = state->csstate.ss.ps.state;
 
 	if (initial_rt_indexes == NIL)
 		return;
@@ -481,10 +642,24 @@ adjust_ri_clauses(ChunkAppendState *state, List *initial_rt_indexes)
 	{
 		Scan *scan = chunk_append_get_scan_plan(lfirst(lc_plan));
 		Index initial_index = lfirst_oid(lc_relid);
+		List *relation_constraints = NIL;
 
-		if (scan != NULL && scan->scanrelid > 0 && scan->scanrelid != initial_index)
+		if (scan != NULL && scan->scanrelid > 0)
 		{
-			ChangeVarNodes(lfirst(lc_clauses), initial_index, scan->scanrelid, 0);
+			Index rt_index = scan->scanrelid;
+			RangeTblEntry *rte = rt_fetch(rt_index, estate->es_range_table);
+			relation_constraints = ca_get_relation_constraints(rte->relid, rt_index, true);
+
+			/*
+			 * Adjust the RangeTableEntry indexes in the restrictinfo
+			 * clauses because during planning subquery indexes may be
+			 * different from the final index after flattening.
+			 */
+			if (rt_index != initial_index)
+				ChangeVarNodes(lfirst(lc_clauses), initial_index, scan->scanrelid, 0);
 		}
+		constraints = lappend(constraints, relation_constraints);
 	}
+	state->initial_constraints = constraints;
+	state->filtered_constraints = constraints;
 }