+++ src/chunk_append/exec.h
@@ -26,19 +26,28 @@ typedef struct ChunkAppendState
 
 	/* list of subplans after planning */
 	List *initial_subplans;
-	/* list of restrictinfo clauses indexed similar to initial_subplans */
+	/* list of constraints indexed like initial_subplans */
+	List *initial_constraints;
+	/* list of restrictinfo clauses indexed like initial_subplans */
 	List *initial_ri_clauses;
 
 	/* list of subplans after startup exclusion */
 	List *filtered_subplans;
+	/* list of relation constraints after startup exclusion */
+	List *filtered_constraints;
 	/* list of restrictinfo clauses after startup exclusion */
 	List *filtered_ri_clauses;
 
 	/* valid subplans for runtime exclusion */
 	Bitmapset *valid_subplans;
+	Bitmapset *params;
 
-	/* sort options if this append is ordered */
+	/* sort options if this append is ordered, only used for EXPLAIN */
 	List *sort_options;
+
+	/* number of loops and exclusions for EXPLAIN */
+	int runtime_number_loops;
+	int runtime_number_exclusions;
 
 } ChunkAppendState;